pi-zero/requirements.txt
pyserial==3.5
requests==2.32.3
RPi.GPIO==0.7.1